added tag cloud

+++ wwwroot/js/Collection.js
@@ -1,12 +1,15 @@
-﻿function setup() {}
-var editurl = document.getElementById('EditUrl').value;
+﻿function setup() { }
+try {
+    var editurl = document.getElementById('EditUrl').value;
+} catch {}
 var viewurl = document.getElementById('ViewUrl').value;
 var defaultimg = "https://www.smartdatajob.com/images/joomlart/demo/default.jpg";
     
 class Collection { 
-    constructor(id, title, topic, description, urlimg) {
+    constructor(id, title, topic, description, urlimg,flag) {
         let container = select('#collections');
-        this.carddiv = createDiv('').addClass('card').parent(container).id(id);
+        this.coldiv = createDiv('').addClass('col').parent(container);
+        this.carddiv = createDiv('').addClass('card').parent(this.coldiv).id(id);
         this.img = createImg(urlimg).addClass('card-img-top imgcard').parent(this.carddiv);
         this.img.elt.addEventListener('error', () => {
             this.img.elt.src = defaultimg;
@@ -14,8 +17,8 @@ class Collection {
         this.cardBody = createDiv('').addClass('card-body').parent(this.carddiv);
         this.fillcardbody(title, topic, description);
         this.cardFooter = createDiv('').addClass('card-footer').parent(this.carddiv);
-        this.fillcardfooter(id);
-        this.addbuttonevents();
+        this.fillcardfooter(id,flag);
+        this.addbuttonevents(flag);
     }
 
     fillcardbody(title, topic, description) {
@@ -24,34 +27,38 @@ class Collection {
         this.description = createElement('p').addClass('card-text').html(description).parent(this.cardBody);
     }
 
-    fillcardfooter(id) {
+    fillcardfooter(id,flag) {
         this.buttonv = createA('#', 'View').addClass('btn btn-primary').parent(this.cardFooter);
-        this.buttone = createA('#', 'Edit').addClass('btn btn-warning').parent(this.cardFooter);
-        this.buttonr = createA('#', 'Remove').addClass('btn btn-secondary').parent(this.cardFooter);
+        if (flag) {
+            this.buttone = createA('#', 'Edit').addClass('btn btn-warning').parent(this.cardFooter);
+            this.buttonr = createA('#', 'Remove').addClass('btn btn-secondary').parent(this.cardFooter);
+        }
     }
 
-    addbuttonevents() {
+    addbuttonevents(flag) {
         this.buttonv.elt.addEventListener('click', () => {
             location.href = viewurl.replace('__id__', this.carddiv.id());
         })
-        this.buttone.elt.addEventListener('click', () => {
-            location.href = editurl.replace('__id__', this.carddiv.id());
-        })
-        this.buttonr.elt.addEventListener('click', () => {
-            var id = this.carddiv.id();
-            var object = this.elt;
-            $.ajax({
-                url: 'Collection/Delete/',
-                type: 'DELETE',
-                data: { 'id': id },
-                success: function (result) {
-                    $(result).remove()
-                }
+        if (flag) {
+            this.buttone.elt.addEventListener('click', () => {
+                location.href = editurl.replace('__id__', this.carddiv.id());
             })
-        })
+            this.buttonr.elt.addEventListener('click', () => {
+                var id = this.carddiv.id();
+                var object = this.elt;
+                $.ajax({
+                    url: 'Collection/Delete/',
+                    type: 'DELETE',
+                    data: { 'id': id },
+                    success: function (result) {
+                        $(result).remove()
+                    }
+                })
+            })
+        }
     }
 }
 
-function CreateCollCard(id, title, topic, description, urlimg) {
-    let coll = new Collection(id, title, topic, description, urlimg);
+function CreateCollCard(id, title, topic, description, urlimg,flag) {
+    let coll = new Collection(id, title, topic, description, urlimg,flag);
 }
